IWTF-525 IWTF-526 IWTF-532 - Grilse loader rework, expose grilse data to reporting feeds

+++ src/main/java/uk/gov/defra/datareturns/data/model/grilse/GrilseWeightGateRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import uk.gov.defra.datareturns.data.BaseRepository;
 
 @RepositoryRestResource
-public interface GrilseWeightGateRepository extends BaseRepository<GrilseWeightGate, Short> {
+public interface GrilseWeightGateRepository extends BaseRepository<GrilseWeightGate, Long> {
 
 }
